[ANDROSDK-393] Remove DataValueHandler interface
