+++ core/src/main/java/org/hisp/dhis/android/core/datavalue/DataValueEndpointCall.java
@@ -74,7 +74,7 @@ public final class DataValueEndpointCall {
         protected CallProcessor<DataValue> processor(GenericCallData data, DataValueQuery query) {
 
             return new TransactionalNoResourceSyncCallProcessor<>(data.databaseAdapter(),
-                    DataValueHandlerImpl.create(data.databaseAdapter()));
+                    DataValueHandler.create(data.databaseAdapter()));
         }
     };
 }
